+++ tasty-plutus/CHANGELOG.md
@@ -4,6 +4,11 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
 ## Unreleased
 
+### Changed
+
+* Rename `paysSelf` and `paysOther` into `paysToSelf` and `paysToOther` for
+  consistency.
+
 ### Removed
 
 * Various re-exports from `Test.Tasty.Plutus.Script.Unit`.